Ensure file state for zip tests is always clean
The files were not necessarily being torn down correctly after the test completes - specifically, the teardown was in the `it` function after an `assert` so would not tear down if the assert failed.

This meant that the file state commencing the test run would not necessarily be clean.

Moving the cleanup into both a `before` and `after` step means that we can be certain that the tests are always executing on a clean and known fs state.